checker/requirements.txt
aiohttp
python-dateutil
kafka-python
pytest
pytest-asyncio
aioresponses